refactor: extract encryption subsystem from read.go into encrypt.go
Move 12 encryption functions, cbcReader struct, passwordPad and
ErrInvalidPassword vars (~265 lines) from read.go into a new encrypt.go.
read.go shrinks from 1246 to 981 lines and drops its four crypto imports
(crypto/aes, crypto/cipher, crypto/md5, crypto/rc4).

Same-package file split — no export changes, no API surface changes.
The cross-file callsite lex.go:512 (decryptString called from readObject)
resolves correctly since both files are in package pdf.
